fix(ui): address PR review comments for MyPrCard
- Extract timeAgo to shared utility src/lib/timeAgo.ts with week/year support
- Fix isMergeable to require pr.state === "open" (excludes closed PRs)
- Use fake timers in test for deterministic time-ago assertion
