+++ src/components/ReviewQueue/ReviewCard.tsx
@@ -1,4 +1,5 @@
 import type { ReactElement } from "react";
+import { timeAgo } from "../../lib/timeAgo";
 import type { PullRequestWithReview } from "../../lib/types";
 import { CI } from "../atoms/CI";
 import { Diff } from "../atoms/Diff";
@@ -9,20 +10,6 @@ interface ReviewCardProps {
   readonly data: PullRequestWithReview;
   readonly onOpen: (url: string) => void;
   readonly onWorkspaceAction?: (workspaceId: string) => void;
-}
-
-function timeAgo(dateStr: string): string {
-  const date = new Date(dateStr);
-  if (Number.isNaN(date.getTime())) return "";
-  const seconds = Math.floor((Date.now() - date.getTime()) / 1000);
-  if (seconds < 0) return "now";
-  if (seconds < 60) return `${seconds}s`;
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes}m`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours}h`;
-  const days = Math.floor(hours / 24);
-  return `${days}d`;
 }
 
 export function ReviewCard({
